Skeleton for Flask Project
I added a main.py file from where the html files will be called. A views.py file which is going to be later used to make interconnections between the html files. A settings.py file which is also still empty but will be needed later. I put the html files in template accordingly to the flask project skeleton.